requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
scikit-learn==1.5.2
streamlit==1.38.0
python-multipart==0.0.9
